LayoutTests/js/property-iteration-expected.txt

This tests ordering of static properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS arrayEqual(before, after) is true
PASS successfullyParsed is true

TEST COMPLETE
